requirement.txt
numpy
ipython
tqdm
matplotlib.pyplot
math
